fix: Styles conforming to system default
Components like Table in ChakraUI implicitly use the system theme
default. However, our code still doesn't support switching themes and
most components are built in light mode. Because of that if the browser
running the website uses dark mode, then the table will be in dark mode
while the rest of the components doesn't.

This fixes it by setting the defaultTheme to light.

Fixes #13